split perf.lua
