+++ spec/helpers/perf.lua
@@ -1,200 +1,12 @@
 local pl_tablex = require("pl.tablex")
-local ngx_pipe = require("ngx.pipe")
-local tty = require("kong.cmd.utils.tty")
-local ffi = require("ffi")
 
-local colors
+local logger = require("spec.helpers.perf.logger")
+local utils = require("spec.helpers.perf.utils")
+local git = require("spec.helpers.perf.git")
 
-if not tty.isatty then
-  colors = setmetatable({}, {__index = function() return "" end})
-else
-  colors = { green = '\27[32m', yellow = '\27[33m', red = '\27[31m', reset = '\27[0m' }
-end
+local my_logger = logger.new_logger("[controller]")
 
-local LOG_LEVEL = ngx.NOTICE
--- how many times for each "driver" operation
 local RETRY_COUNT = 3
-local DRIVER
-
--- Some logging helpers
-local level_cfg = {
-  [ngx.DEBUG] = { "debug", colors.green },
-  [ngx.INFO] = { "info", "" },
-  [ngx.NOTICE] = { "notice", "" },
-  [ngx.WARN] = { "warn",  colors.yellow },
-  [ngx.ERR] = { "error", colors.red },
-  [ngx.CRIT] = { "crit", colors.red },
-}
-
-local function set_log_level(lvl)
-  if not level_cfg[lvl] then
-    error("Unknown log level ", lvl, 2)
-  end
-  LOG_LEVEL = lvl
-end
-
-local function log(lvl, namespace, ...)
-  lvl = lvl or ngx.INFO
-  local lvl_literal, lvl_color = unpack(level_cfg[lvl] or {"info", ""})
-  if lvl <= LOG_LEVEL then
-    ngx.update_time()
-    local msec = ngx.now()
-    print(lvl_color,
-          ("%s%s [%s] %s "):format(
-            ngx.localtime():sub(12),
-            ("%.3f"):format(msec - math.floor(msec)):sub(2),
-            lvl_literal, namespace
-          ),
-          table.concat({...}, ""),
-          colors.reset)
-  end
-end
-local function new_logger(namespace)
-  return setmetatable({
-    debug = function(...) log(ngx.DEBUG, namespace, ...) end,
-    info = function(...) log(ngx.INFO, namespace, ...) end,
-    warn = function(...) log(ngx.WARN, namespace, ...) end,
-    err = function(...) log(ngx.ERR, namespace, ...) end,
-    crit = function(...) log(ngx.CRIT, namespace, ...) end,
-    log_exec = function(...) log(ngx.DEBUG, namespace, "=> ", ...) end,
-  }, {
-    __call = function(self, lvl, ...) log(lvl, namespace, ...) end,
-  })
-end
-local my_logger = new_logger("[controller]")
-
-string.startswith = function(s, start) -- luacheck: ignore
-  return s and start and start ~= "" and s:sub(1, #start) == start
-end
-
-string.endswith = function(s, e) -- luacheck: ignore
-  return s and e and e ~= "" and s:sub(#s-#e+1, #s) == e
-end
-
---- Spawns a child process and get its exit code and outputs
--- @param opts.stdin string the stdin buffer
--- @param opts.logger function(lvl, _, line) stdout+stderr writer; if not defined, whole
--- stdoud and stderr is returned
--- @param opts.stop_signal function return true to abort execution
--- @return stdout+stderr, err if opts.logger not set; bool+err if opts.logger set
-local function execute(cmd, opts)
-  -- my_logger.debug("exec: ", cmd)
-
-  local proc, err = ngx_pipe.spawn(cmd, {
-    merge_stderr = true,
-  })
-  if not proc then
-    return false, "failed to start process: " .. err
-  end
-
-  -- set stdout/stderr read timeout to 1s for faster noticing process exit
-  -- proc:set_timeouts(write_timeout?, stdout_read_timeout?, stderr_read_timeout?, wait_timeout?)
-  proc:set_timeouts(nil, 1000, 1000, nil)
-  if opts and opts.stdin then
-    proc:write(opts.stdin)
-  end
-  proc:shutdown("stdin")
-
-  local log_output = opts and opts.logger
-  local ret = {}
-
-  while true do
-    -- is it alive?
-    local ok = proc:kill(0)
-    if not ok then
-      break
-    end
-
-    local l, err = proc:stdout_read_line()
-    if l then
-      if log_output then
-        opts.logger(l)
-      else
-        table.insert(ret, l)
-      end
-    end
-    if err == "closed" then
-      break
-    end
-    local sig = opts and opts.stop_signal and opts.stop_signal()
-    if sig then
-      proc:kill(sig)
-      break
-    end
-  end
-  local ok, msg, code = proc:wait()
-  ok = ok and code == 0
-  ret = log_output and ok or table.concat(ret, "\n")
-  if ok then
-    return ret
-  else
-    return ret, ("process exited with code %d: %s"):format(code, msg)
-  end
-end
-
---- Execute a command and return until pattern is found in its output
--- @function wait_output
--- @param cmd string the command the execute
--- @param pattern string the pattern to find in stdout and stderr
--- @param timeout number time in seconds to wait for the pattern
--- @return bool whether the pattern is found
-local function wait_output(cmd, pattern, timeout)
-  timeout = timeout or 5
-  local found
-  local co = coroutine.create(function()
-    while not found do
-      local line = coroutine.yield("yield")
-      if line:match(pattern) then
-        found = true
-      end
-    end
-  end)
-
-  -- start
-  coroutine.resume(co)
-
-  -- don't kill it, it me finish by itself
-  ngx.thread.spawn(function()
-    execute(cmd, {
-      logger = function(line)
-        return coroutine.running(co) and coroutine.resume(co, line)
-      end,
-      stop_signal = function() if found then return 9 end end,
-    })
-  end)
-
-  ngx.update_time()
-  local s = ngx.now()
-  while not found and ngx.now() - s <= timeout do
-    ngx.update_time()
-    ngx.sleep(0.1)
-  end
-
-  return found
-end
-
-ffi.cdef [[
-  int setenv(const char *name, const char *value, int overwrite);
-  int unsetenv(const char *name);
-]]
-
---- Set an environment variable
--- @function setenv
--- @param env (string) name of the environment variable
--- @param value the value to set
--- @return true on success, false otherwise
-local function setenv(env, value)
-  return ffi.C.setenv(env, value, 1) == 0
-end
-
-
---- Unset an environment variable
--- @function setenv
--- @param env (string) name of the environment variable
--- @return true on success, false otherwise
-local function unsetenv(env)
-  return ffi.C.unsetenv(env) == 0
-end
 
 -- Real user facing functions
 local driver_functions = {
@@ -277,13 +89,19 @@ end
 
 local _M = {
   use_driver = use_driver,
-  new_logger = new_logger,
-  set_log_level = set_log_level,
-  setenv = setenv,
-  unsetenv = unsetenv,
   set_retry_count = set_retry_count,
-  execute = execute,
-  wait_output = wait_output,
+
+  new_logger = logger.new_logger,
+  set_log_level = logger.set_log_level,
+
+  setenv = utils.setenv,
+  unsetenv = utils.unsetenv,
+  execute = utils.execute,
+  wait_output = utils.wait_output,
+
+  git_checkout = git.git_checkout,
+  git_restore = git.git_restore,
+  get_kong_version = git.get_kong_version,
 }
 
 --- Start the upstream (nginx) with given conf
@@ -359,7 +177,7 @@ function _M.start_load(opts)
   load_should_stop = false
 
   load_thread = ngx.thread.spawn(function()
-    return execute(load_cmd,
+    return utils.execute(load_cmd,
         {
           stop_signal = function() if load_should_stop then return 9 end end,
         })
@@ -383,14 +201,14 @@ function _M.start_stapxx(sample_name, ...)
   stapxx_should_stop = false
 
   stapxx_thread = ngx.thread.spawn(function()
-    return execute(start_cmd,
+    return utils.execute(start_cmd,
         {
           stop_signal = function() if stapxx_should_stop then return 3 end end,
         })
   end)
 
   local wait_cmd = invoke_driver("get_wait_stapxx_cmd")
-  if not wait_output(wait_cmd, "stap_", 30) then
+  if not utils.wait_output(wait_cmd, "stap_", 30) then
     return false, "timeout waiting systemtap probe to load"
   end
 
@@ -511,8 +329,6 @@ function _M.wait_stap_probe(timeout)
   return invoke_driver("wait_stap_probe", timeout or 20)
 end
 
-local git_stashed, git_head
-
 --- Generate the flamegraph and return SVG
 -- @function generate_flamegraph
 -- @return Nothing. Throws an error if any.
@@ -556,73 +372,6 @@ function _M.save_error_log(filename)
   invoke_driver("save_error_log", filename)
 
   my_logger.debug("Kong error log written to ", filename)
-end
-
-function _M.git_checkout(version)
-  if not execute("which git") then
-    error("git binary not found")
-  end
-
-  local res, err
-  local hash, _ = execute("git rev-parse HEAD")
-  if not hash or not hash:match("[a-f0-f]+") then
-    error("Unable to parse HEAD pointer, is this a git repository?")
-  end
-
-  -- am i on a named branch/tag?
-  local n, _ = execute("git rev-parse --abbrev-ref HEAD")
-  if n and n ~= "HEAD" then
-    hash = n
-  end
-  -- anything to save?
-  n, err = execute("git status --untracked-files=no --porcelain")
-  if not err and (n and #n > 0) then
-    my_logger.info("saving your working directory")
-    res, err = execute("git stash save kong-perf-test-autosaved")
-    if err then
-      error("Cannot save your working directory: " .. err .. (res or "nil"))
-    end
-    git_stashed = true
-  end
-
-  my_logger.debug("switching away from ", hash, " to ", version)
-
-  res, err = execute("git checkout " .. version)
-  if err then
-    error("Cannot switch to " .. version .. ":\n" .. res)
-  end
-  if not git_head then
-    git_head = hash
-  end
-end
-
-function _M.git_restore()
-  if git_head then
-    local res, err = execute("git checkout " .. git_head)
-    if err then
-      return false, "git checkout: " .. res
-    end
-    git_head = nil
-
-    if git_stashed then
-      local res, err = execute("git stash pop")
-      if err then
-        return false, "git stash pop: " .. res
-      end
-      git_stashed = false
-    end
-  end
-end
-
-function _M.get_kong_version()
-  -- unload the module if it's previously loaded
-  package.loaded["kong.meta"] = nil
-
-  local ok, meta, _ = pcall(require, "kong.meta")
-  if ok then
-    return meta._VERSION
-  end
-  error("can't read Kong version from kong.meta: " .. (meta or "nil"))
 end
 
 return _M
